--- app/helpers/batch_invitation_permissions_helper.rb
@@ -1,9 +1,0 @@
-module BatchInvitationPermissionsHelper
-  def formatted_permission_name(application_name, permission_name)
-    if permission_name == SupportedPermission::SIGNIN_NAME
-      "Has access to #{application_name}?"
-    else
-      permission_name
-    end
-  end
-end
