+++ app/helpers/users_helper.rb
@@ -88,4 +88,12 @@ module UsersHelper
   def supported_permission_checkbox_id(application, permission)
     "user_application_#{application.id}_supported_permission_#{permission.id}"
   end
+
+  def formatted_permission_name(application_name, permission_name)
+    if permission_name == SupportedPermission::SIGNIN_NAME
+      "Has access to #{application_name}?"
+    else
+      permission_name
+    end
+  end
 end
